requirements.txt

fastapi
uvicorn
pymongo
voyageai
python-dotenv
requests
llama-index
llama-index-embeddings-fireworks
llama-index-llms-fireworks
